Restructure keyupdate test code
